requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
asyncpg
greenlet
python-multipart
